Added routeParser.py
This takes the route.kml in our kml directory, parses it, uses the Google
Elevation API to get elevation data and uses matplotlib to create a png image
of the plot of this data.

With some extensions, this could be part of a post preprocessor if we decide to
take that route.