Refactored and adding variable sizing
Split routeParser up into a few functions so that a preprocessor could use it
in the near future. These functions are also documented!

Made it so that the size could be set for the elevation plot which as always
with matplotlib involved some hackyass code. Maybe it would be better if we
used the google charts api to make the elevation profiles...